Split RepositoryTests into 3 classes
