--- Tests/Krecha.Lib.Tests/Data/RepositoryTests.cs
@@ -1,699 +1,0 @@
-﻿using AutoFixture;
-using Krecha.Lib.Data;
-using Krecha.Lib.Data.Models;
-using Krecha.Lib.Tests.Helpers;
-using Microsoft.EntityFrameworkCore;
-
-namespace Krecha.Lib.Tests.Data;
-public class RepositoryTests
-{
-    private readonly Repository<Currency, SettlementsDbContext> _currencyRepository;
-    private readonly Repository<Settlement, SettlementsDbContext> _settlementRepository;
-    private readonly Repository<SettlementEntry, SettlementsDbContext> _settlementEntryRepository;
-    private readonly SettlementsDbContext _dbContext;
-    private readonly Fixture _fixture = new();
-
-    public RepositoryTests()
-    {
-        _dbContext = EFHelpers.SetupInMemoryDbContext();
-        _currencyRepository = new(_dbContext);
-        _settlementRepository = new(_dbContext);
-        _settlementEntryRepository = new(_dbContext);
-    }
-
-    [Fact]
-    public async Task GetAll_ForCurrencies_ShouldReturnAllItemsInDb()
-    {
-        // Arrange
-        int expectedCurrencyCount = 10;
-        var currencies = CreateAndAddToDbTestCurrencies(expectedCurrencyCount);
-
-        // Act
-        var actualCurrencies = _currencyRepository.GetAll();
-        int actualCurrencyCount = await actualCurrencies.CountAsync();
-       
-        // Assert
-        Assert.Equal(expectedCurrencyCount, actualCurrencyCount);
-        
-        foreach(var expected in currencies)
-        {
-            var actual = await actualCurrencies.FirstOrDefaultAsync(currency => currency.Id == expected.Id);
-
-            Assert.NotNull(actual);
-            Assert.Equivalent(expected, actual);
-        }
-    }
-
-    [Fact]
-    public async Task GetAll_ForSettlements_ShouldReturnAllItemsInDb()
-    {
-        // Arrange
-        int expectedSettlementCount = 10;
-        var settlements = CreateAndAddToDbTestSettlements(expectedSettlementCount);
-
-        // Act
-        var actualSettlements = _settlementRepository.GetAll();
-        int actualSettlementCount = await actualSettlements.CountAsync();
-
-        // Assert
-        Assert.Equal(expectedSettlementCount, actualSettlementCount);
-
-        foreach (var expected in settlements)
-        {
-            var actual = await actualSettlements.FirstOrDefaultAsync(settlement => settlement.Id == expected.Id);
-
-            Assert.NotNull(actual);
-            Assert.Equivalent(expected, actual);
-        }
-    }
-    
-    [Fact]
-    public async Task GetAll_ForSettlementEntries_ShouldReturnAllItemsInDb()
-    {
-        // Arrange
-        int expectedSettlementEntryCount = 10;
-        var settlementEntries = CreateAndAddToDbTestSettlementEntries(expectedSettlementEntryCount);
-
-        // Act
-        var actualSettlementEntries = _settlementEntryRepository.GetAll();
-        int actualSettlementEntryCount = await actualSettlementEntries.CountAsync();
-
-        // Assert
-        Assert.Equal(expectedSettlementEntryCount, actualSettlementEntryCount);
-
-        foreach (var expected in settlementEntries)
-        {
-            var actual = await actualSettlementEntries.FirstOrDefaultAsync(entry => entry.Id == expected.Id);
-
-            Assert.NotNull(actual);
-            Assert.Equivalent(expected, actual);
-        }
-    }
-
-    [Fact]
-    public async Task GetById_WhenCurrencyDoesntExist_ShouldReturnNull()
-    {
-        // Arrange
-        int currencyId = 1;
-
-        // Act
-        var actualCurrency = await _currencyRepository.GetById(currencyId);
-
-        // Assert
-        Assert.Null(actualCurrency);
-    }
-    
-    [Fact]
-    public async Task GetById_WhenSettlementDoesntExist_ShouldReturnNull()
-    {
-        // Arrange
-        int settlementId = 2;
-
-        // Act
-        var actualSettlement = await _settlementRepository.GetById(settlementId);
-
-        // Assert
-        Assert.Null(actualSettlement);
-    }
-    
-    [Fact]
-    public async Task GetById_WhenSettlementEntryDoesntExist_ShouldReturnNull()
-    {
-        // Arrange
-        int settlementEntryId = 3;
-
-        // Act
-        var actualSettlementEntry = await _settlementEntryRepository.GetById(settlementEntryId);
-
-        // Assert
-    }
-
-    [Fact]
-    public async Task GetById_WhenCurrencyExists_ShouldReturnIt()
-    {
-        // Arrage
-        Currency expected = CreateAndAddToDbTestCurrencies(1).First();
-
-        // Act
-        var actual = await _currencyRepository.GetById(expected.Id);
-
-        // Assert
-        Assert.NotNull(actual);
-        Assert.Equivalent(expected, actual);
-    }
-    
-    [Fact]
-    public async Task GetById_WhenSettlementExists_ShouldReturnIt()
-    {
-        // Arrage
-        Settlement expected = CreateAndAddToDbTestSettlements(1).First();
-
-        // Act
-        var actual = await _settlementRepository.GetById(expected.Id);
-
-        // Assert
-        Assert.NotNull(actual);
-        Assert.Equivalent(expected, actual);
-    }
-    
-    [Fact]
-    public async Task GetById_WhenSettlementEntryExists_ShouldReturnIt()
-    {
-        // Arrage
-        SettlementEntry expected = CreateAndAddToDbTestSettlementEntries(1).First();
-
-        // Act
-        var actual = await _settlementEntryRepository.GetById(expected.Id);
-
-        // Assert
-        Assert.NotNull(actual);
-        Assert.Equivalent(expected, actual);
-    }
-
-    [Fact]
-    public async Task Create_ForCurrency_ShouldReturnTheSameObject()
-    {
-        // Arrange
-        Currency expected = CreateTestCurrencies(1).First();
-
-        // Act
-        var actual = await _currencyRepository.Create(expected);
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-    
-    [Fact]
-    public async Task Create_ForSettlement_ShouldReturnTheSameObject()
-    {
-        // Arrange
-        Settlement expected = CreateTestSettlements(1).First();
-
-        // Act
-        var actual = await _settlementRepository.Create(expected);
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-    
-    [Fact]
-    public async Task Create_ForSettlementEntry_ShouldReturnTheSameObject()
-    {
-        // Arrange
-        SettlementEntry expected = CreateTestSettlementEntries(1).First();
-
-        // Act
-        var actual = await _settlementEntryRepository.Create(expected);
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-
-    [Fact]
-    public async Task Create_ForCurrency_ShouldAddItToDb()
-    {
-        // Arrange
-        Currency expected = CreateTestCurrencies(1).First();
-        var expectedEntityState = EntityState.Unchanged;
-
-        // Act
-        await _currencyRepository.Create(expected);
-        var actual = await _dbContext.Currencies.FindAsync(expected.Id);
-        var actuaEntityState = _dbContext.Entry(expected).State;
-
-        // Assert
-        Assert.NotNull(actual);
-        Assert.Equal(expected, actual);
-        Assert.Equal(expectedEntityState, actuaEntityState);
-    }
-    
-    [Fact]
-    public async Task Create_ForSettlement_ShouldAddItToDb()
-    {
-        // Arrange
-        Settlement expected = CreateTestSettlements(1).First();
-        var expectedEntityState = EntityState.Unchanged;
-
-        // Act
-        await _settlementRepository.Create(expected);
-        var actual = await _dbContext.Settlements.FindAsync(expected.Id);
-        var actuaEntityState = _dbContext.Entry(expected).State;
-
-        // Assert
-        Assert.NotNull(actual);
-        Assert.Equal(expected, actual);
-        Assert.Equal(expectedEntityState, actuaEntityState);
-    }
-
-    [Fact]
-    public async Task Create_ForSettlementEntry_ShouldAddItToDb()
-    {
-        // Arrange
-        SettlementEntry expected = CreateTestSettlementEntries(1).First();
-        var expectedEntityState = EntityState.Unchanged;
-
-        // Act
-        await _settlementEntryRepository.Create(expected);
-        var actual = await _dbContext.Entries.FindAsync(expected.Id);
-        var actuaEntityState = _dbContext.Entry(expected).State;
-
-        // Assert
-        Assert.NotNull(actual);
-        Assert.Equal(expected, actual);
-        Assert.Equal(expectedEntityState, actuaEntityState);
-    }
-
-    [Fact]
-    public async Task Create_ForCurrencyThatExists_ShouldThrowArgumentException()
-    {
-        // Arrange
-        Currency currencyInDb = CreateAndAddToDbTestCurrencies(1).First();
-
-        // Act
-        var act = () => _currencyRepository.Create(currencyInDb);
-
-        // Assert
-        await Assert.ThrowsAnyAsync<ArgumentException>(act);
-    }
-    
-    [Fact]
-    public async Task Create_ForSettlementThatExists_ShouldThrowArgumentException()
-    {
-        // Arrange
-        Settlement settlementInDb = CreateAndAddToDbTestSettlements(1).First();
-
-        // Act
-        var act = () => _settlementRepository.Create(settlementInDb);
-
-        // Assert
-        await Assert.ThrowsAnyAsync<ArgumentException>(act);
-    }
-    
-    [Fact]
-    public async Task Create_ForSettlementEntryThatExists_ShouldThrowArgumentException()
-    {
-        // Arrange
-        SettlementEntry entryInDb = CreateAndAddToDbTestSettlementEntries(1).First();
-
-        // Act
-        var act = () => _settlementEntryRepository.Create(entryInDb);
-
-        // Assert
-        await Assert.ThrowsAnyAsync<ArgumentException>(act);
-    }
-
-    [Fact]
-    public async Task Update_WhenCurrencyDoesntExist_ShouldReturnNull()
-    {
-        // Arrange
-        int currencyId = 11;
-
-        // Act
-        var actual = await _currencyRepository.Update(currencyId, currency => { });
-
-        // Assert
-        Assert.Null(actual);
-    }
-    
-    [Fact]
-    public async Task Update_WhenSettlementDoesntExist_ShouldReturnNull()
-    {
-        // Arrange
-        int settlementId = 11;
-
-        // Act
-        var actual = await _settlementRepository.Update(settlementId, settlement => { });
-
-        // Assert
-        Assert.Null(actual);
-    }
-    
-    [Fact]
-    public async Task Update_WhenSettlementEntryDoesntExist_ShouldReturnNull()
-    {
-        // Arrange
-        int entryId = 11;
-
-        // Act
-        var actual = await _settlementEntryRepository.Update(entryId, entry => { });
-
-        // Assert
-        Assert.Null(actual);
-    }
-
-    [Fact]
-    public async Task Update_WhenCurrencyExists_ShouldApplyAllChanges()
-    {
-        // Arrange
-        Currency currency = CreateAndAddToDbTestCurrencies(1).First();
-        string expectedName = "Test Name";
-        string expectedSymbol = "Test Symbol";
-        CurrencySymbolPosition expectedSymbolPosition = CurrencySymbolPosition.After;
-
-        // Act
-        var actual = await _currencyRepository.Update(currency.Id, currency =>
-        {
-            currency.Name = expectedName;
-            currency.Symbol = expectedSymbol;
-            currency.SymbolPosition = expectedSymbolPosition;
-        });
-
-        // Assert
-        Assert.NotNull(actual);
-        Assert.Equal(expectedName, actual.Name);
-        Assert.Equal(expectedSymbol, actual.Symbol);
-        Assert.Equal(expectedSymbolPosition, actual.SymbolPosition);
-    }
-
-    [Fact]
-    public async Task Update_WhenSettlementExists_ShouldApplyAllChanges()
-    {
-        // Arrange
-        Settlement settlement = CreateAndAddToDbTestSettlements(1).First();
-        string expectedName = "Test Name";
-        string expectedDescription = "Test Description";
-        bool expectedArchived = true;
-
-        // Act
-        var actual = await _settlementRepository.Update(settlement.Id, settlement =>
-        {
-            settlement.Name = expectedName;
-            settlement.Description = expectedDescription;
-            settlement.Archived = expectedArchived;
-        });
-
-        // Assert
-        Assert.NotNull(actual);
-        Assert.Equal(expectedName, actual.Name);
-        Assert.Equal(expectedDescription, actual.Description);
-        Assert.Equal(expectedArchived, actual.Archived);
-    }
-    
-    [Fact]
-    public async Task Update_WhenSettlementEntryExists_ShouldApplyAllChanges()
-    {
-        // Arrange
-        SettlementEntry entry = CreateAndAddToDbTestSettlementEntries(1).First();
-        DateTime expectedCreatedDate = DateTime.Now.AddMinutes(-19);
-        string expectedDescription = "Test Description";
-        decimal expectedAmount = 1138;
-        bool expectedArchived = true;
-
-        // Act
-        var actual = await _settlementEntryRepository.Update(entry.Id, entry =>
-        {
-            entry.CreatedDate = expectedCreatedDate;
-            entry.Description = expectedDescription;
-            entry.Amount = expectedAmount;
-            entry.Archived = expectedArchived;
-        });
-
-        // Assert
-        Assert.NotNull(actual);
-        Assert.Equal(expectedCreatedDate, actual.CreatedDate);
-        Assert.Equal(expectedDescription, actual.Description);
-        Assert.Equal(expectedAmount, actual.Amount);
-        Assert.Equal(expectedArchived, actual.Archived);
-    }
-    
-    [Fact]
-    public async Task Update_WhenCurrencyExists_ShouldReturnTheSameObject()
-    {
-        // Arrange
-        Currency expected = CreateAndAddToDbTestCurrencies(1).First();
-
-        // Act
-        var actual = await _currencyRepository.Update(expected.Id, currency => { });
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-
-    [Fact]
-    public async Task Update_WhenSettlementExists_ShouldReturnTheSameObject()
-    {
-        // Arrange
-        Settlement expected = CreateAndAddToDbTestSettlements(1).First();
-
-        // Act
-        var actual = await _settlementRepository.Update(expected.Id, currency => { });
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-    
-    [Fact]
-    public async Task Update_WhenSettlementEntryExists_ShouldReturnTheSameObject()
-    {
-        // Arrange
-        SettlementEntry expected = CreateAndAddToDbTestSettlementEntries(1).First();
-
-        // Act
-        var actual = await _settlementEntryRepository.Update(expected.Id, currency => { });
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-
-    [Fact]
-    public async Task Update_WhenCurrencyExists_ShouldResultInCorrectEntityState()
-    {
-        // Arrange
-        Currency currency = CreateAndAddToDbTestCurrencies(1).First();
-        var expected = EntityState.Unchanged;
-
-        // Act
-        var result = await _currencyRepository.Update(currency.Id, currency => { });
-        var actual = _dbContext.Entry(result!).State;
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-    
-    [Fact]
-    public async Task Update_WhenSettlementExists_ShouldResultInCorrectEntityState()
-    {
-        // Arrange
-        Settlement settlement = CreateAndAddToDbTestSettlements(1).First();
-        var expected = EntityState.Unchanged;
-
-        // Act
-        var result = await _settlementRepository.Update(settlement.Id, currency => { });
-        var actual = _dbContext.Entry(result!).State;
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-    
-    [Fact]
-    public async Task Update_WhenSettlementEntryExists_ShouldResultInCorrectEntityState()
-    {
-        // Arrange
-        SettlementEntry entry = CreateAndAddToDbTestSettlementEntries(1).First();
-        var expected = EntityState.Unchanged;
-
-        // Act
-        var result = await _settlementEntryRepository.Update(entry.Id, currency => { });
-        var actual = _dbContext.Entry(result!).State;
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-
-    [Fact]
-    public async Task Delete_WhenCurrencyDoesntExist_ShouldReturnNull()
-    {
-        // Arrange
-        int currencyId = 19;
-
-        // Act
-        var actual = await _currencyRepository.Delete(currencyId);
-
-        // Assert
-        Assert.Null(actual);
-    }
-    
-    [Fact]
-    public async Task Delete_WhenSettlementDoesntExist_ShouldReturnNull()
-    {
-        // Arrange
-        int settlementId = 19;
-
-        // Act
-        var actual = await _settlementRepository.Delete(settlementId);
-
-        // Assert
-        Assert.Null(actual);
-    }
-    
-    [Fact]
-    public async Task Delete_WhenSettlementEntryDoesntExist_ShouldReturnNull()
-    {
-        // Arrange
-        int settlementEntryId = 19;
-
-        // Act
-        var actual = await _settlementEntryRepository.Delete(settlementEntryId);
-
-        // Assert
-        Assert.Null(actual);
-    }
-
-    [Fact]
-    public async Task Delete_WhenCurrencyExists_ShouldReturnTheSameObject()
-    {
-        // Arrange
-        Currency expected = CreateAndAddToDbTestCurrencies(1).First();
-
-        // Act
-        var actual = await _currencyRepository.Delete(expected.Id);
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-    
-    [Fact]
-    public async Task Delete_WhenSettlementExists_ShouldReturnTheSameObject()
-    {
-        // Arrange
-        Settlement expected = CreateAndAddToDbTestSettlements(1).First();
-
-        // Act
-        var actual = await _settlementRepository.Delete(expected.Id);
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-    
-    [Fact]
-    public async Task Delete_WhenSettlementEntryExists_ShouldReturnTheSameObject()
-    {
-        // Arrange
-        SettlementEntry expected = CreateAndAddToDbTestSettlementEntries(1).First();
-
-        // Act
-        var actual = await _settlementEntryRepository.Delete(expected.Id);
-
-        // Assert
-        Assert.Equal(expected, actual);
-    }
-
-    [Fact]
-    public async Task Delete_WhenCurrencyExists_ShouldRemoveItFromDb()
-    {
-        // Arrange
-        Currency currency = CreateAndAddToDbTestCurrencies(1).First();
-        var expectedEntityState = EntityState.Detached;
-
-        // Act
-        await _currencyRepository.Delete(currency.Id);
-        var actual = await _dbContext.Currencies.FindAsync(currency.Id);
-        var actualEntityState = _dbContext.Entry(currency).State;
-
-        // Assert
-        Assert.Null(actual);
-        Assert.Equal(expectedEntityState, actualEntityState);
-    }
-
-    [Fact]
-    public async Task Delete_WhenSettlementExists_ShouldRemoveItFromDb()
-    {
-        // Arrange
-        Settlement settlement = CreateAndAddToDbTestSettlements(1).First();
-        var expectedEntityState = EntityState.Detached;
-
-        // Act
-        await _settlementRepository.Delete(settlement.Id);
-        var actual = await _dbContext.Settlements.FindAsync(settlement.Id);
-        var actualEntityState = _dbContext.Entry(settlement).State;
-
-        // Assert
-        Assert.Null(actual);
-        Assert.Equal(expectedEntityState, actualEntityState);
-    }
-    
-    [Fact]
-    public async Task Delete_WhenSettlementEntryExists_ShouldRemoveItFromDb()
-    {
-        // Arrange
-        SettlementEntry entry = CreateAndAddToDbTestSettlementEntries(1).First();
-        var expectedEntityState = EntityState.Detached;
-
-        // Act
-        await _settlementEntryRepository.Delete(entry.Id);
-        var actual = await _dbContext.Entries.FindAsync(entry.Id);
-        var actualEntityState = _dbContext.Entry(entry).State;
-
-        // Assert
-        Assert.Null(actual);
-        Assert.Equal(expectedEntityState, actualEntityState);
-    }
-
-    private List<Currency> CreateAndAddToDbTestCurrencies(int count)
-    {
-        var currencies = CreateTestCurrencies(count);
-        AddEntitiesToInMemoryDb(currencies);
-        return currencies;
-    }
-
-    private List<Settlement> CreateAndAddToDbTestSettlements(int count)
-    {
-        var settlements = CreateTestSettlements(count);
-        AddEntitiesToInMemoryDb(settlements);
-        return settlements;
-    }
-
-    private List<SettlementEntry> CreateAndAddToDbTestSettlementEntries(int count)
-    {
-        var entries = CreateTestSettlementEntries(count);
-        AddEntitiesToInMemoryDb(entries);
-        return entries;
-    }
-
-    private List<Currency> CreateTestCurrencies(int count)
-    {
-        List<Currency> output = _fixture
-            .Build<Currency>()
-                .Without(currency => currency.Id)
-                .Without(currency => currency.Settlements)
-            .CreateMany(count)
-            .ToList();
-
-        return output;
-    }
-    
-    private List<Settlement> CreateTestSettlements(int count)
-    {
-        List<Settlement> output = _fixture
-            .Build<Settlement>()
-                .Without(settlement => settlement.Id)
-                .Without(settlement => settlement.Entries)
-                .Without(settlement => settlement.Currency)
-            .CreateMany(count)
-            .ToList();
-
-        return output;
-    }
-    
-    private List<SettlementEntry> CreateTestSettlementEntries(int count)
-    {
-        List<SettlementEntry> output = _fixture
-            .Build<SettlementEntry>()
-                .Without(entry => entry.Id)
-                .Without(entry => entry.Settlement)
-            .CreateMany(count)
-            .ToList();
-
-        return output;
-    }
-
-    private void AddEntitiesToInMemoryDb<TEntity>(List<TEntity> entities)
-        where TEntity : class
-    {
-        _dbContext
-            .Set<TEntity>()
-            .AddRange(entities);
-
-        _dbContext.SaveChanges();
-    }
-}
